+++ test/hephaestus/steps/done_test.exs
@@ -9,7 +9,7 @@ defmodule Hephaestus.Steps.DoneTest do
 
   describe "execute/3" do
     test "returns done event" do
-      instance = Instance.new(TestWorkflow, %{})
+      instance = Instance.new(TestWorkflow, 1, %{})
       context = Context.new(%{order_id: 123})
 
       result = Done.execute(instance, nil, context)
@@ -18,7 +18,7 @@ defmodule Hephaestus.Steps.DoneTest do
     end
 
     test "ignores config" do
-      instance = Instance.new(TestWorkflow, %{})
+      instance = Instance.new(TestWorkflow, 1, %{})
       context = Context.new(%{})
 
       result = Done.execute(instance, %{some: "config"}, context)